feat[纹理]：增加轨道

+++ sample/src/main/java/com/jiangpengyong/sample/utils/MathExt.kt
@@ -9,5 +9,5 @@ fun Float.toRadians(): Double {
 }
 
 fun Double.toRadians(): Double {
-    return Math.toRadians(this.toDouble())
+    return Math.toRadians(this)
 }
